# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Novate Platform API",
        "description": "Enterprise Digital Workspace Platform API. This API provides endpoints for authentication, organization management, dynamic form building, data collection, project management, analytics, reports, and notifications.",
        "contact": {
            "name": "Novate Platform Support",
            "email": "support@novate.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/exports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "List exports",
                "description": "Returns paginated list of report exports for the current organization.",
                "operationId": "listExports",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exports retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create export",
                "description": "Queues an export job for submissions in CSV, XLSX, or PDF format.",
                "operationId": "createExport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "submissions"
                                    },
                                    "format": {
                                        "type": "string",
                                        "example": "csv"
                                    },
                                    "filters": {
                                        "properties": {
                                            "form_uuid": {
                                                "type": "string",
                                                "format": "uuid",
                                                "nullable": true
                                            },
                                            "status": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Export queued"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/exports/{reportExport}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get export",
                "description": "Returns a single export record with its status and file path.",
                "operationId": "getExport",
                "parameters": [
                    {
                        "name": "reportExport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "Enter your Bearer token obtained from auth/login or auth/register."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}